deleted scene that does not work
going to create a main scene to work from, and my scene which will inlcude levels ideas that I will try to come up with. this is good cause if someone wanted to use the mechanics they can use the mechanics in the main scene, and make their own scene so they can make their own level.